app/requirements.txt
blinker==1.8.2
certifi>=2024.07.04
cffi==1.16.0
charset-normalizer==3.3.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
numpy==1.26.4
pillow==10.3.0
exifread==3.0.0
pycparser==2.22
python-dotenv==1.0.1
rawpy==0.21.0
requests==2.32.4
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3==2.6.3
Werkzeug==3.1.5
opencv-contrib-python-headless==4.10.0.84
scikit-image==0.22.0
pytest==6.2.4